# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Power Vital API",
    "version": "1.0.0",
    "description": "Power Vital e-commerce + MLM platform backend.\n\n## Languages\nAll user-facing strings live in three locales: **TR** (primary), **RU**, **KG**.\nThe active locale is negotiated via the `Accept-Language` header and a `lang`\nquery param. AI-generated translations fill missing locale fields automatically.\n\n## Auth\nMutating endpoints require a `Bearer` JWT in the `Authorization` header.\nAdmin-only routes additionally check the `role` claim.\n\n## Rate limiting\nAll `/api/*` routes are rate-limited to 150 req/min per IP. Burst over\nthe limit returns 429 with `Retry-After`.\n\n## Errors\nAll errors return `{ error: string, issues?: ValidationIssue[] }` where\n`ValidationIssue` has the shape `{ path, message, code }`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://powervital.kg",
      "description": "Production (Kyrgyzstan)"
    },
    {
      "url": "https://powervital.org",
      "description": "Production (intl alias)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from POST /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/system/config": {
      "get": {
        "tags": [
          "System"
        ],
        "description": "Fetch system config + revenue/bonus/payout stats (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Config + stats"
          }
        }
      },
      "put": {
        "tags": [
          "System"
        ],
        "description": "Update system config (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isMlmEnabled": {
                    "type": "boolean"
                  },
                  "maxPayoutLimitPct": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "isFastStartActive": {
                    "type": "boolean"
                  },
                  "fastStartRates": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\[[\\d,\\s]+\\]$"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 0
                        }
                      }
                    ]
                  },
                  "isUnilevelActive": {
                    "type": "boolean"
                  },
                  "unilevelRates": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\[[\\d,\\s]+\\]$"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 0
                        }
                      }
                    ]
                  },
                  "isOverdriveActive": {
                    "type": "boolean"
                  },
                  "overdrivePoolPct": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      }
    },
    "/api/v1/system/close-week": {
      "post": {
        "tags": [
          "System"
        ],
        "description": "Close the current week and roll points over (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Week closed"
          }
        }
      }
    },
...
  },
...
}